move editors back to one file
