--- packages/datocms-plugin-table/src/components/ContentEditor/style.module.css
@@ -1,16 +1,0 @@
-.editor {
-  position: relative;
-}
-
-.editorType {
-  position: absolute;
-  right: 0;
-  height: 38px;
-  padding: 0 8px;
-  color: #adaaaa;
-  font-size: 12px;
-  z-index: 5;
-  border-left: 1px solid #adaaaa;
-  display: flex;
-  align-items: center;
-}
