+++ packages/datocms-plugin-table/src/components/EditableCell/style.module.css
@@ -48,7 +48,3 @@
   width: 24px;
   height: 24px;
 }
-
-.textEditor {
-  min-width: 100%;
-}
