+++ packages/datocms-plugin-table/src/components/TextEditor/style.module.css
@@ -1,3 +1,7 @@
+.textEditor {
+  min-width: 100%;
+}
+
 .editor {
   position: relative;
 }
